+++ activestorage/lib/active_storage/previewer.rb
@@ -1,14 +1,12 @@
 # frozen_string_literal: true
 
-require "active_storage/downloading"
+require "active_support/core_ext/string/filters"
 
 module ActiveStorage
   # This is an abstract base class for previewers, which generate images from blobs. See
   # ActiveStorage::Previewer::PDFPreviewer and ActiveStorage::Previewer::VideoPreviewer for examples of
   # concrete subclasses.
   class Previewer
-    include Downloading
-
     attr_reader :blob
 
     # Implement this method in a concrete subclass. Have it return true when given a blob from which
@@ -28,6 +26,11 @@ module ActiveStorage
     end
 
     private
+      # Downloads the blob to a tempfile on disk. Yields the tempfile.
+      def download_blob_to_tempfile #:doc:
+        blob.open(&block)
+      end
+
       # Executes a system command, capturing its binary output in a tempfile. Yields the tempfile.
       #
       # Use this method to shell out to a system library (e.g. mupdf or ffmpeg) for preview image